chore: reorganize document
