+++ src/common/components/Header.tsx
@@ -2,7 +2,7 @@ import Image from "next/legacy/image";
 import Link from "next/link";
 import logo from "public/images/logo.svg";
 import login from "public/images/login.svg";
-import logout from "public/images/logout.svg";
+import logout from "public/images/header/logout.svg";
 import StudentDropdown from "@/modules/dashboard/student/StudentDropdown";
 import NutritionistDropdown from "@/modules/dashboard/nutritionist/NutritionistDropdown";
 import { useEffect, useState } from "react";
