+++ src/common/components/dietary-record/bodyRateChart/BodyRate.tsx
