+++ src/common/components/dietary-record/bodyRateChart/BodyRateChart.tsx
