+++ src/common/components/dietary-record/courseInfo/CourseInfo.tsx
