+++ src/modules/dashboard/student/info/StudentInfoForm.tsx
@@ -4,7 +4,7 @@ import {
   commonErrMsgClass,
   commonPhonePattern,
   commonRequiredErrMsg,
-} from "@/common/lib/dashboard/errMsg/commonErrMsg";
+} from "@/common/lib/errMsg/commonErrMsg";
 import { getCookie } from "cookies-next";
 import { useProfilePutApiMutation } from "@/common/redux/service/profile";
 import useEditForm from "@/common/hooks/useEditForm";
